Update of simulation configurator and generator.. there was a problem with int overflow, therefore there is a new update to int->double

+++ src/manualGenerator.cc
@@ -42,7 +42,7 @@ void ManualGenerator::initialize()
     }
 }
 
-void ManualGenerator::updateLambda(int bandwidth){
+void ManualGenerator::updateLambda(double bandwidth){
     lambda= bandwidth / length/8;
 }
